+++ Cargo.toml
@@ -15,5 +15,5 @@ members = [
     "juniper_warp",
     "juniper_actix",
     "tests/codegen",
-    "tests/integration"
+    "tests/integration",
 ]
