+++ tests/integration/Cargo.toml
@@ -5,13 +5,13 @@ edition = "2021"
 publish = false
 
 [dev-dependencies]
+async-trait = "0.1.39"
 chrono = "0.4"
 derive_more = "0.99"
+fnv = "1.0"
 futures = "0.3"
 juniper = { path = "../../juniper" }
 juniper_subscriptions = { path = "../../juniper_subscriptions" }
-async-trait = "0.1.39"
-fnv = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 tokio = { version = "1.0", features = ["rt", "macros", "time"] }
