+++ tests/integration/tests/codegen_derive_object_with_raw_idents.rs
