+++ tests/integration/tests/codegen_scalar_attr_derive_input.rs
@@ -1,15 +1,18 @@
-//! Tests for `#[derive(GraphQLScalar)]` macro.
+//! Tests for `#[graphql_scalar]` macro placed on [`DeriveInput`].
+//!
+//! [`DeriveInput`]: syn::DeriveInput
+
 pub mod common;
 
 use std::fmt;
 
 use chrono::{DateTime, TimeZone, Utc};
 use juniper::{
-    execute, graphql_object, graphql_value, graphql_vars, GraphQLScalar, InputValue,
+    execute, graphql_object, graphql_scalar, graphql_value, graphql_vars, InputValue,
     ParseScalarResult, ParseScalarValue, ScalarToken, ScalarValue, Value,
 };
 
-use common::{
+use self::common::{
     util::{schema, schema_with_scalar},
     MyScalarValue,
 };
@@ -17,7 +20,7 @@ use common::{
 mod trivial {
     use super::*;
 
-    #[derive(GraphQLScalar)]
+    #[graphql_scalar]
     struct Counter(i32);
 
     impl Counter {
@@ -93,8 +96,7 @@ mod trivial {
 mod transparent {
     use super::*;
 
-    #[derive(GraphQLScalar)]
-    #[graphql(transparent)]
+    #[graphql_scalar(transparent)]
     struct Counter(i32);
 
     struct QueryRoot;
@@ -154,8 +156,10 @@ mod transparent {
 mod transparent_with_resolver {
     use super::*;
 
-    #[derive(GraphQLScalar)]
-    #[graphql(transparent, to_output_with = Self::to_output)]
+    #[graphql_scalar(
+        transparent,
+        to_output_with = Self::to_output,
+    )]
     struct Counter(i32);
 
     impl Counter {
@@ -221,12 +225,11 @@ mod transparent_with_resolver {
 mod all_custom_resolvers {
     use super::*;
 
-    #[derive(GraphQLScalar)]
-    #[graphql(
+    #[graphql_scalar(
         to_output_with = to_output,
         from_input_with = from_input,
     )]
-    #[graphql(parse_token_with = parse_token)]
+    #[graphql_scalar(parse_token_with = parse_token)]
     struct Counter(i32);
 
     fn to_output<S: ScalarValue>(v: &Counter) -> Value<S> {
@@ -300,8 +303,7 @@ mod all_custom_resolvers {
 mod explicit_name {
     use super::*;
 
-    #[derive(GraphQLScalar)]
-    #[graphql(name = "Counter")]
+    #[graphql_scalar(name = "Counter")]
     struct CustomCounter(i32);
 
     impl CustomCounter {
@@ -377,8 +379,7 @@ mod explicit_name {
 mod delegated_parse_token {
     use super::*;
 
-    #[derive(GraphQLScalar)]
-    #[graphql(parse_token(i32))]
+    #[graphql_scalar(parse_token(i32))]
     struct Counter(i32);
 
     impl Counter {
@@ -450,8 +451,7 @@ mod delegated_parse_token {
 mod multiple_delegated_parse_token {
     use super::*;
 
-    #[derive(GraphQLScalar)]
-    #[graphql(parse_token(String, i32))]
+    #[graphql_scalar(parse_token(String, i32))]
     enum StringOrInt {
         String(String),
         Int(i32),
@@ -510,8 +510,7 @@ mod multiple_delegated_parse_token {
 mod where_attribute {
     use super::*;
 
-    #[derive(GraphQLScalar)]
-    #[graphql(
+    #[graphql_scalar(
         to_output_with = to_output,
         from_input_with = from_input,
         parse_token(String),
@@ -591,8 +590,7 @@ mod where_attribute {
 mod with_self {
     use super::*;
 
-    #[derive(GraphQLScalar)]
-    #[graphql(with = Self)]
+    #[graphql_scalar(with = Self)]
     struct Counter(i32);
 
     impl Counter {
@@ -668,8 +666,7 @@ mod with_self {
 mod with_module {
     use super::*;
 
-    #[derive(GraphQLScalar)]
-    #[graphql(
+    #[graphql_scalar(
         with = custom_date_time,
         parse_token(String),
         where(Tz: From<Utc>, Tz::Offset: fmt::Display),
@@ -753,8 +750,7 @@ mod description_from_doc_comment {
     use super::*;
 
     /// Description
-    #[derive(GraphQLScalar)]
-    #[graphql(parse_token(i32))]
+    #[graphql_scalar(parse_token(i32))]
     struct Counter(i32);
 
     impl Counter {
@@ -830,8 +826,7 @@ mod description_from_attribute {
     use super::*;
 
     /// Doc comment
-    #[derive(GraphQLScalar)]
-    #[graphql(description = "Description from attribute", parse_token(i32))]
+    #[graphql_scalar(description = "Description from attribute", parse_token(i32))]
     struct Counter(i32);
 
     impl Counter {
@@ -907,8 +902,7 @@ mod custom_scalar {
     use super::*;
 
     /// Description
-    #[derive(GraphQLScalar)]
-    #[graphql(scalar = MyScalarValue, parse_token(i32))]
+    #[graphql_scalar(scalar = MyScalarValue, parse_token(i32))]
     struct Counter(i32);
 
     impl Counter {
@@ -984,8 +978,7 @@ mod generic_scalar {
     use super::*;
 
     /// Description
-    #[derive(GraphQLScalar)]
-    #[graphql(scalar = S: ScalarValue, parse_token(i32))]
+    #[graphql_scalar(scalar = S: ScalarValue, parse_token(i32))]
     struct Counter(i32);
 
     impl Counter {
@@ -1051,7 +1044,7 @@ mod generic_scalar {
             execute(DOC, None, &schema, &graphql_vars! {}, &()).await,
             Ok((
                 graphql_value!({"__type": {"description": "Description"}}),
-                vec![],
+                vec![]
             )),
         );
     }
@@ -1060,8 +1053,7 @@ mod generic_scalar {
 mod bounded_generic_scalar {
     use super::*;
 
-    #[derive(GraphQLScalar)]
-    #[graphql(scalar = S: ScalarValue + Clone, parse_token(i32))]
+    #[graphql_scalar(scalar = S: ScalarValue + Clone, parse_token(i32))]
     struct Counter(i32);
 
     impl Counter {
@@ -1072,7 +1064,7 @@ mod bounded_generic_scalar {
         fn from_input<S: ScalarValue>(v: &InputValue<S>) -> Result<Self, String> {
             v.as_int_value()
                 .map(Self)
-                .ok_or_else(|| format!("Expected `String`, found: {v}"))
+                .ok_or_else(|| format!("Expected `Counter`, found: {v}"))
         }
     }
