+++ tests/integration/tests/codegen_scalar_value_derive.rs
