+++ tests/integration/tests/codegen_union_attr.rs
@@ -1,10 +1,13 @@
 //! Tests for `#[graphql_union]` macro.
+
 pub mod common;
 
 use juniper::{
     execute, graphql_object, graphql_union, graphql_value, graphql_vars, DefaultScalarValue,
-    EmptyMutation, EmptySubscription, GraphQLObject, GraphQLType, RootNode, ScalarValue,
+    GraphQLObject, ScalarValue,
 };
+
+use self::common::util::{schema, schema_with_scalar};
 
 #[derive(GraphQLObject)]
 struct Human {
@@ -50,31 +53,6 @@ pub struct DroidCustomContext {
 struct EwokCustomContext {
     id: String,
     funny: bool,
-}
-
-fn schema<'q, C, Q>(query_root: Q) -> RootNode<'q, Q, EmptyMutation<C>, EmptySubscription<C>>
-where
-    Q: GraphQLType<DefaultScalarValue, Context = C, TypeInfo = ()> + 'q,
-{
-    RootNode::new(
-        query_root,
-        EmptyMutation::<C>::new(),
-        EmptySubscription::<C>::new(),
-    )
-}
-
-fn schema_with_scalar<'q, S, C, Q>(
-    query_root: Q,
-) -> RootNode<'q, Q, EmptyMutation<C>, EmptySubscription<C>, S>
-where
-    Q: GraphQLType<S, Context = C, TypeInfo = ()> + 'q,
-    S: ScalarValue + 'q,
-{
-    RootNode::new_with_scalar_value(
-        query_root,
-        EmptyMutation::<C>::new(),
-        EmptySubscription::<C>::new(),
-    )
 }
 
 mod trivial {
@@ -571,8 +549,9 @@ mod explicit_scalar {
 }
 
 mod custom_scalar {
+    use crate::common::MyScalarValue;
+
     use super::*;
-    use common::MyScalarValue;
 
     #[graphql_union(scalar = MyScalarValue)]
     trait Character {
